status_detail: replace legacy StatusDetailFragment with new one
