+++ app/src/main/kotlin/io/github/anthonyeef/cattle/databinding/StatusItemClickEventHandler.kt
@@ -1,13 +1,11 @@
 package io.github.anthonyeef.cattle.databinding
 
 import io.github.anthonyeef.cattle.activity.PhotoDisplayActivity
-import io.github.anthonyeef.cattle.activity.StatusDetailActivity
-import io.github.anthonyeef.cattle.activity.StatusDetailActivity.Companion.EXTRA_STATUS_HAS_CONVERSATION
-import io.github.anthonyeef.cattle.activity.StatusDetailActivity.Companion.EXTRA_STATUS_ID
 import io.github.anthonyeef.cattle.constant.app
 import io.github.anthonyeef.cattle.data.statusData.Status
 import io.github.anthonyeef.cattle.extension.show
 import io.github.anthonyeef.cattle.fragment.ProfileFragment
+import io.github.anthonyeef.cattle.fragment.StatusDetailFragment
 import org.jetbrains.anko.intentFor
 import org.jetbrains.anko.newTask
 import org.jetbrains.anko.support.v4.withArguments
@@ -17,7 +15,9 @@ import org.jetbrains.anko.support.v4.withArguments
  */
 object StatusItemClickEventHandler {
     fun onClickStatus(status: Status) {
-        app.startActivity(app.intentFor<StatusDetailActivity>(EXTRA_STATUS_ID to status.id, EXTRA_STATUS_HAS_CONVERSATION to status.inReplyToStatusId.isNotEmpty()).newTask())
+      StatusDetailFragment()
+          .withArguments(StatusDetailFragment.KEY_STATUS_ID to status.id)
+          .show()
     }
 
     fun openProfile(status: Status) {
